fix(miner): accommodate for an empty BusTree and undefined block number
This fix accommodates the scenario where the smart contract returns 0x00
for an empty BusTree. This adjustment is necessary as the smart
contract's design leverages this return value to save gas during
deployment. As a result, this patch updates the miner to handle this
case appropriately when attempting to mine the first batch.

Secondly, if there are no batches mined yet, the subgraph returns an
undefined block number from which the EventScanner should begin work.
This patch also addresses this issue, by using new env variable of the
genesis block.