feat(miner): remove UTXO count and reward from SC call signature
This commit removes the ability for miners to specify the number of
UTXOs and rewards in the smart contract call. Instead, the smart
contract now defines these values. Additionally, recent testing has
shown that RPC nodes cannot reliably scan more than 100 blocks at a
time, so the number of blocks to scan has been limited to 100.