feat(miner): update miner to accommodate new BusTree smart contracts
The BusTree smart contracts introduced new parameters `maturityBlock`
and `remainingBlocks` to prevent miners from mining empty queues.

This commit includes changes in the miner to work with this new
functionality in the smart contracts. Specifically, the miner now
considers the `remainingBlocks` to decide whether a queue can be
processed or not.