fix(miner): replace skip argument with orderBy and first
An error occurred when making a request to the subgraph: "the skip
argument must be between 0 and 5000, but is 6000". It seems that we
cannot use the skip argument with more than 5000 elements. Large skip
values generally perform poorly. Therefore, when retrieving a large
number of items, it's more effective to page through entities based on
an attribute.

This patch fixes this issue by using `orderBy` and `first`
instead of the `skip` argument.