feat(sdk): force simulation option to control UTXO simulation
This commit introduces a new environment variable FORCE_UTXO_SIMULATION
in the sdk. This new variable allows control over the UTXO simulation
process in the miner. If set to 'true', the miner will always add a
simulated UTXO, regardless of whether there are any pending queues. If
set to 'false', a simulated UTXO will only be added when a pending queue
exists.

The addition of this variable provides more flexibility in controlling
the mining process and can be used to optimize the performance of the
miner under different conditions.