fix(sdk): Correctly validate pending queue existence
The previous implementation of the `hasPendingQueues` method in the SDK
always returned true, due to the fact that pending queues are always
present in the smart contract, even when they contain no UTXOs. This led
to an incorrect evaluation of whether a simulation was needed in the
`doWork` function.

This commit modifies the `hasPendingQueues` method to explicitly check
for the presence of UTXOs in the queue, thus ensuring accurate results.
Additionally, we have refactored the simulation need check in `doWork`
into a separate `checkSimulationNeed` function for better readability
and maintainability.

The changes are important to prevent unnecessary simulations when there
are no UTXOs in the queue, improving the system's efficiency.