fix(sdk): Transition from Business Source License to MIT License
This commit updates the license from Business Source License 1.1 to MIT
License. The change was necessary to ensure broader compatibility and
easier adoption of the SDK by the community. All SPDX-License-Identifier
and SPDX-FileCopyrightText headers in the source files have been updated
to reflect this change. The license text in the LICENSE file has also
been replaced with the standard MIT License text. This transition aligns
with our commitment to open source software and its principles.